requirements.txt
# Core functionality
python-dotenv==1.1.0
pydantic==2.11.5

# LangChain and AI
langchain-google-genai==2.1.4
langchain-core==0.3.61
langsmith==0.3.42

# Data processing
pandas==2.2.3
pillow==11.2.1
unidecode==1.4.0
fuzzywuzzy==0.18.0
python-levenshtein==0.27.1

# Google APIs
gspread==6.2.1
google-auth-oauthlib==1.2.2
google-auth==2.40.2
google-api-python-client==2.170.0

# Dependencies for the above packages
requests==2.32.3
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
urllib3==2.4.0
cachetools==5.5.2
pyasn1==0.6.1
pyasn1-modules==0.4.2
rsa==4.9.1
oauthlib==3.2.2
requests-oauthlib==2.0.0
httplib2==0.22.0
googleapis-common-protos==1.70.0
protobuf==6.31.0
google-api-core==2.24.2
google-auth-httplib2==0.2.0
uritemplate==4.1.1
pydantic-core==2.33.2
rapidfuzz==3.13.0
jsonpatch==1.33
jsonpointer==3.0.0
orjson==3.10.18
tenacity==9.1.2
packaging==24.2
pyyaml==6.0.2
typing-extensions==4.13.2
